Cleaned up some unused imports and reverted changes to Search button position
