+++ test/USE_CASE/FilterUseCaseTest.java
@@ -1,13 +1,13 @@
-package INTERFACE_ADAPTER;
+package USE_CASE;
 
 import DATA_ACCESS.DatabaseDAO;
 import ENTITY.Event;
 import ENTITY.EventFactory;
 import ENTITY.EventFactoryInterface;
 import ENTITY.User;
-import VIEW.SearchEventsView;
+import VIEW.FilterEventsView;
+import VIEW_CREATOR.FilterEventsViewFactory;
 import VIEW_CREATOR.LoadMapViewModel;
-import VIEW_CREATOR.SearchEventsViewFactory;
 import org.jxmapviewer.JXMapKit;
 import org.jxmapviewer.viewer.WaypointPainter;
 
@@ -24,7 +24,7 @@ import java.util.Set;
 
 import static org.junit.Assert.assertNotNull;
 
-public class SearchUseCaseTest {
+public class FilterUseCaseTest {
     static String message;
     static boolean popUpDiscovered;
     public Event addEvent() {
@@ -32,28 +32,25 @@ public class SearchUseCaseTest {
         DatabaseDAO databaseDAO = new DatabaseDAO();
         User newUser = databaseDAO.getUserByUsername("ff");
         List<User> attendees = new ArrayList<>();
-        Event event = eventFactory.create("TestEvent2", "Food and Drinks", "Example Event", LocalDate.now(), LocalTime.now(), newUser, attendees, 43.645306, -79.380247);
-        Event event2 = eventFactory.create("TestEvent1", "Food and Drinks", "Example Event", LocalDate.now(), LocalTime.now(), newUser, attendees, 43.645306, -79.380247);
+        Event event = eventFactory.create("TestEvent", "Food and Drinks", "Example Event", LocalDate.now(), LocalTime.now(), newUser, attendees, 43.645306, -79.380247);
 
         databaseDAO.addEvent(event);
-        databaseDAO.addEvent(event2);
         return event;
     }
 
-    public SearchEventsView MakeSearchEventsView(){
+    public FilterEventsView MakeFilterEventsView(){
         LoadMapViewModel viewModel = new LoadMapViewModel();
-        SearchEventsViewFactory searchEventsViewFactory = new SearchEventsViewFactory();
-        SearchEventsView searchEventsView = searchEventsViewFactory.create(viewModel);
-        searchEventsView.setVisible(false);
-        return searchEventsView;
+        FilterEventsViewFactory filterEventsViewFactory = new FilterEventsViewFactory();
+        FilterEventsView filterEventsView = filterEventsViewFactory.create(viewModel);
+        return filterEventsView;
     }
 
     @org.junit.Test
-    public void testEventShownAfterSearch() {
+    public void testEventShownAfterFilter() {
         Event event = addEvent();
 
         DatabaseDAO databaseDAO = new DatabaseDAO();
-        Set<Event> allEvents = databaseDAO.SearchEvent("");
+        Set<Event> allEvents = databaseDAO.FilterEvents("");
         int event_id = -1;
         for(Event event2 : allEvents) {
             if (event2.getEventName().equals(event.getEventName())){
@@ -61,16 +58,16 @@ public class SearchUseCaseTest {
             }
         }
 
-        SearchEventsView searchEventsView = MakeSearchEventsView();
-        searchEventsView.getController().setEvents(allEvents);
+        FilterEventsView filterEventsView = MakeFilterEventsView();
+        filterEventsView.getController().setEvents(allEvents);
 
-        JTextField searchTextField = getTextField();
-        searchTextField.setText("TestEvent");
+        JRadioButton foodButton = getFilterButton(3);
+        foodButton.setSelected(true);
 
         JButton applyButton = getButton(1);
         applyButton.doClick();
 
-        JXMapKit mapKit = searchEventsView.getViewModel().getMapKit();
+        JXMapKit mapKit = filterEventsView.getViewModel().getMapKit();
 
         Set<Event> eventsShown = ((WaypointPainter<Event>)mapKit.getMainMap().getOverlayPainter()).getWaypoints();
 
@@ -85,11 +82,11 @@ public class SearchUseCaseTest {
     }
 
     @org.junit.Test
-    public void testEventNotShownAfterSearch() {
+    public void testEventNotShownAfterFilter() {
         Event event = addEvent();
-        DatabaseDAO databaseDAO = new DatabaseDAO();
-        Set<Event> allEvents = databaseDAO.SearchEvent("");
 
+        DatabaseDAO databaseDAO = new DatabaseDAO();
+        Set<Event> allEvents = databaseDAO.FilterEvents("");
         int event_id = -1;
         for(Event event2 : allEvents) {
             if (event2.getEventName().equals(event.getEventName())){
@@ -97,18 +94,19 @@ public class SearchUseCaseTest {
             }
         }
 
-        SearchEventsView searchEventsView = MakeSearchEventsView();
-        searchEventsView.getController().setEvents(allEvents);
+        FilterEventsView filterEventsView = MakeFilterEventsView();
+        filterEventsView.getController().setEvents(allEvents);
 
-        JTextField searchTextField = getTextField();
-        searchTextField.setText("TestEvent1");
+        JRadioButton foodButton = getFilterButton(2);
+        foodButton.setSelected(true);
 
         JButton applyButton = getButton(1);
         applyButton.doClick();
 
-        JXMapKit mapKit = searchEventsView.getViewModel().getMapKit();
+        JXMapKit mapKit = filterEventsView.getViewModel().getMapKit();
 
         Set<Event> eventsShown = ((WaypointPainter<Event>)mapKit.getMainMap().getOverlayPainter()).getWaypoints();
+
         Set<Integer> eventsShownID = new HashSet<>();
         for(Event event2 : eventsShown) {
             eventsShownID.add(event2.getEventId());
@@ -120,32 +118,25 @@ public class SearchUseCaseTest {
     }
 
     @org.junit.Test
-    public void testNoEventsFoundPopUp(){
+    public void testShowAllFilter() {
         DatabaseDAO databaseDAO = new DatabaseDAO();
-        Set<Event> allEvents = databaseDAO.SearchEvent("");
+        Set<Event> allEvents = databaseDAO.FilterEvents("");
 
-        SearchEventsView searchEventsView = MakeSearchEventsView();
-        searchEventsView.getController().setEvents(allEvents);
-        createCloseTimer().start();
-        searchEventsView.getController().execute("TestEvent12345677");
+        FilterEventsView filterEventsView = MakeFilterEventsView();
+        filterEventsView.getController().setEvents(allEvents);
 
-        assert(popUpDiscovered);
-    }
+        JRadioButton foodButton = getFilterButton(2);
+        foodButton.setSelected(true);
 
-    @org.junit.Test
-    public void testCancelButtonFunctionality(){
+        JButton applyButton = getButton(1);
+        applyButton.doClick();
 
-        DatabaseDAO databaseDAO = new DatabaseDAO();
-        Set<Event> allEvents = databaseDAO.SearchEvent("");
+        JRadioButton showAllButton = getFilterButton(8);
+        showAllButton.setSelected(true);
 
-        SearchEventsView searchEventsView = MakeSearchEventsView();
-        searchEventsView.getController().setEvents(allEvents);
-        searchEventsView.getController().execute("TestEvent");
-        JButton cancelButton = getButton(0);
-        cancelButton.doClick();
+        applyButton.doClick();
 
-        JXMapKit mapKit = searchEventsView.getViewModel().getMapKit();
-
+        JXMapKit mapKit = filterEventsView.getViewModel().getMapKit();
         Set<Event> eventsShown = ((WaypointPainter<Event>)mapKit.getMainMap().getOverlayPainter()).getWaypoints();
 
         assert (eventsShown.equals(allEvents));
@@ -153,17 +144,36 @@ public class SearchUseCaseTest {
     }
 
     @org.junit.Test
-    public void testEmptyStringApplyButtonPopUp() {
-
-        MakeSearchEventsView();
-        popUpDiscovered = false;
-
+    public void testEmptyApplyButtonPopUp() {
+        MakeFilterEventsView();
         JButton applyButton = getButton(1);
         createCloseTimer().start();
         applyButton.doClick();
 
         assert(popUpDiscovered);
 
+    }
+
+    @org.junit.Test
+    public void testNoEventsFoundPopUp() {
+        DatabaseDAO databaseDAO = new DatabaseDAO();
+        Set<Event> allEvents = databaseDAO.FilterEvents("");
+
+        FilterEventsView filterEventsView = MakeFilterEventsView();
+        filterEventsView.getController().setEvents(allEvents);
+
+        createCloseTimer().start();
+        filterEventsView.getController().execute("NotRealType", filterEventsView.getViewModel());
+        assert(popUpDiscovered);
+    }
+
+    @org.junit.Test
+    public void testCancelButtonDisposal() {
+        FilterEventsView filterEventsView = MakeFilterEventsView();
+        filterEventsView.showMenu();
+        JButton cancelButton = getButton(0);
+        cancelButton.doClick();
+        assert (!filterEventsView.getFrame().isVisible());
     }
     public JFrame getFrame(){
         JFrame mainFrame = null;
@@ -193,7 +203,7 @@ public class SearchUseCaseTest {
 
     }
 
-    public JTextField getTextField(){
+    public JRadioButton getFilterButton(int i) {
         JFrame mainFrame = getFrame();
 
         mainFrame.getComponent(0);
@@ -201,9 +211,11 @@ public class SearchUseCaseTest {
         Component contentPant = ((JRootPane) root).getContentPane();
         JPanel jPanel = (JPanel) contentPant;
         JPanel mainPanel = (JPanel) jPanel.getComponent(0);
-        JTextField jTextField = (JTextField) mainPanel.getComponent(1);
+        JPanel filterButtonsPanel = (JPanel) mainPanel.getComponent(1);
+        JRadioButton button = (JRadioButton) filterButtonsPanel.getComponent(i);
 
-        return jTextField;
+        return button;
+
     }
 
     private Timer createCloseTimer() {
@@ -226,8 +238,8 @@ public class SearchUseCaseTest {
                             System.out.println("message = " + s);
 
                             // store the information we got from the JDialog
-                            SearchUseCaseTest.message = s;
-                            SearchUseCaseTest.popUpDiscovered = true;
+                            FilterUseCaseTest.message = s;
+                            FilterUseCaseTest.popUpDiscovered = true;
 
                             window.dispose();
                         }
@@ -241,5 +253,4 @@ public class SearchUseCaseTest {
         t.setRepeats(false);
         return t;
     }
-
 }
